pyproject.toml
[tool.poetry]
name = "bi360"
version = "0.1.0"
description = ""
authors = ["Brayant Gualdron <zbgualdron@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.13"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2 = "^2.9.9"
django-filter = "^23.3"
pyyaml = "^6.0.1"
djangorestframework-simplejwt = "^5.3.0"
django-cors-headers = "^4.3.0"
django-environ = "^0.11.2"
gunicorn = "^21.2.0"
numpy = "^1.26.1"
pandas = "^2.1.1"
openpyxl = "^3.1.2"
sqlalchemy = "^2.0.22"
msal = "^1.25.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
colorlog = "^6.7.0"

[tool.isort]
multi_line_output = 5
line_length = 119

[tool.yapf]
based_on_style = "google"
align_closing_bracket_with_visual_indent = true
coalesce_brackets = true
column_limit = 119
dedent_closing_brackets = true

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
